Add files via upload
I have made a working plugin for the Konami Code that you can activate at anytime during gameplay, this to activate a control switch (by default Switch 3, because that's what I needed it for). Use it for whatever, doesn't matter.

It is free for anyone to use, commercial or otherwise. No restrictions.

Up, up, down, down, left, right, left, right, B, A (X, Z on keyboard).